+++ src/main/java/me/niloybiswas/spblog/exception/ResourceNotFoundException.java
@@ -11,9 +11,9 @@ public class ResourceNotFoundException extends RuntimeException{
 	
 	String resourceName;
 	String fieldName;
-	BigInteger fieldValue;
+	Long fieldValue;
 	
-	public ResourceNotFoundException(String resourceName, String fieldName, BigInteger fieldValue) {
+	public ResourceNotFoundException(String resourceName, String fieldName, Long fieldValue) {
 		super(String.format("%s not found with %s : %s", resourceName, fieldName, fieldValue));
 		this.resourceName = resourceName;
 		this.fieldName = fieldName;
