+++ src/main/java/me/niloybiswas/spblog/repository/CategoryRepo.java
@@ -9,6 +9,6 @@ import me.niloybiswas.spblog.entitiy.Category;
 
 
 @Repository
-public interface CategoryRepo extends JpaRepository<Category, BigInteger>{
+public interface CategoryRepo extends JpaRepository<Category, Long>{
 
 }
