+++ src/main/java/me/niloybiswas/spblog/repository/UserRepo.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 import me.niloybiswas.spblog.entitiy.User;
 
 @Repository
-public interface UserRepo extends JpaRepository<User, BigInteger>{
+public interface UserRepo extends JpaRepository<User, Long>{
 
 }
