+++ src/main/java/me/niloybiswas/spblog/service/PostService.java
@@ -1,35 +1,31 @@
 package me.niloybiswas.spblog.service;
 
 import me.niloybiswas.spblog.dto.PostDTO;
-import me.niloybiswas.spblog.entitiy.Category;
-import me.niloybiswas.spblog.entitiy.Post;
-import me.niloybiswas.spblog.entitiy.User;
 
-import java.math.BigInteger;
 import java.util.List;
 
 public interface PostService {
 
     // create
-    PostDTO createPost(PostDTO postDTO, BigInteger userId, BigInteger categoryId);
+    PostDTO createPost(PostDTO postDTO, Long userId, Long categoryId);
 
     // update
-    PostDTO updatePost(PostDTO postDTO, BigInteger postId);
+    PostDTO updatePost(PostDTO postDTO, Long postId);
 
     // delete
-    void deletePost(BigInteger postId);
+    void deletePost(Long postId);
 
     // get all
     List<PostDTO> getAllPosts();
 
     // get by id
-    PostDTO getPostById(BigInteger postId);
+    PostDTO getPostById(Long postId);
 
     // get all by category
-    List<PostDTO> getPostsByCategory(BigInteger categoryId);
+    List<PostDTO> getPostsByCategory(Long categoryId);
 
     // get all by user
-    List<PostDTO> getPostsByUser(BigInteger userId);
+    List<PostDTO> getPostsByUser(Long userId);
 
     // search
     List<PostDTO> searchPosts(String keyword);
